+++ backend/src/main/java/com/example/webduck/global/security/encryption/AttributeEncryptor.java
@@ -1,4 +1,4 @@
-package com.example.webduck.config.security.encryption;
+package com.example.webduck.global.security.encryption;
 
 import static com.example.webduck.global.exception.exceptionCode.LogicExceptionCode.BAD_REQUEST;
