+++ backend/src/main/java/com/example/webduck/global/security/oauth/entity/userInfo/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package com.example.webduck.config.security.oauth.entity.userInfo;
+package com.example.webduck.global.security.oauth.entity.userInfo;
 
 import com.example.webduck.member.entity.SocialType;
 import java.util.Map;
