+++ backend/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 spring:
   profiles:
-    active: test # prod,dev,test
+    active: dev # prod,dev,test
 ---
