+++ backend/src/test/java/com/example/webduck/member/customMock/MockMemberUtil.java
@@ -1,6 +1,6 @@
 package com.example.webduck.member.customMock;
 
-import com.example.webduck.config.security.oauth.entity.SessionMember;
+import com.example.webduck.global.security.oauth.entity.SessionMember;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
